backend/requirements.txt
torch>=2.0
transformers>=4.41
huggingface_hub>=0.23
fastapi>=0.110
uvicorn[standard]>=0.29
python-chess>=1.999
pydantic>=2.0
